+++ src/Chronic/Handlers/SdRmnHandler.cs
@@ -10,7 +10,7 @@ namespace Chronic.Handlers
         public Span Handle(IList<Token> tokens, Options options)
         {
             var month = tokens[1].GetTag<RepeaterMonthName>();
-            var day = tokens[0].GetTag<ScalarDay>().Value;
+            var day = (int)tokens[0].GetTag<ScalarDay>().Value;
             if (Time.IsMonthOverflow(options.Clock().Year, (int)month.Value, day))
             {
                 return null;
