+++ src/Chronic/Handlers/SmSdSyHandler.cs
@@ -10,8 +10,8 @@ namespace Chronic.Handlers
         public virtual Span Handle(IList<Token> tokens, Options options)
         {
             var month = (int)tokens[0].GetTag<ScalarMonth>().Value;
-            var day = tokens[1].GetTag<ScalarDay>().Value;
-            var year = tokens[2].GetTag<ScalarYear>().Value;
+            var day = (int)tokens[1].GetTag<ScalarDay>().Value;
+            var year = (int)tokens[2].GetTag<ScalarYear>().Value;
 
             var time_tokens = tokens.Skip(3).ToList();
             if (Time.IsMonthOverflow(year, month, day))
