+++ src/Chronic/Tags/Repeaters/RepeaterHour.cs
@@ -80,10 +80,10 @@ namespace Chronic.Tags.Repeaters
             return new Span(hourStart, hourEnd);
         }
 
-        public override Span GetOffset(Span span, int amount, Pointer.Type pointer)
+        public override Span GetOffset(Span span, decimal amount, Pointer.Type pointer)
         {
             var direction = (pointer == Pointer.Type.Future) ? 1 : -1;
-            return span.Add(direction * amount * RepeaterHour.HOUR_SECONDS);
+            return span.Add((long)(direction * amount * RepeaterHour.HOUR_SECONDS));
         }
     }
 }
